// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  checkouts: typeof checkouts;
  credits: typeof credits;
  emailTemplates: typeof emailTemplates;
  feedback: typeof feedback;
  http: typeof http;
  polar: typeof polar;
  rateLimit: typeof rateLimit;
  resend: typeof resend;
  users: typeof users;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;